break authServiceFactory.ts into smaller bits
