+++ .changeset/sour-cooks-cough.md
@@ -1,0 +1,6 @@
+---
+'@backstage/backend-app-api': patch
+'@backstage/backend-common': patch
+---
+
+Minor internal restructuring
