+++ packages/backend-common/src/auth/createLegacyAuthAdapters.ts
@@ -36,7 +36,7 @@ import {
   createCredentialsWithUserPrincipal,
   createCredentialsWithNonePrincipal,
   toInternalBackstageCredentials,
-} from '../../../backend-app-api/src/services/implementations/auth/authServiceFactory';
+} from '../../../backend-app-api/src/services/implementations/auth/helpers';
 // TODO is this circular thingy a problem? Test in e2e
 import {
   type IdentityApiGetIdentityRequest,
